Fix duplicating buff icons
This could happen if the player had a (de)buff on the ignore list. The issue didn't exist previously because we were preemptively wiping the (de)buff list on every pass, but now we're only wiping extra entries as an optimization. The problem here was that we were counting all buffs, instead of just displayed ones.